Trim the variable resolution and its tests
`varSource` distinguished six sources where resolution only asks four
questions: three of the constants were assigned and never compared. Collapse
them into `sourceSupplied`, which is what the code means by them.

`parseString` wrote the captured `val` and returned an error, so the value
flowed two ways at once; it returns the value instead.

The nine tfcompat tests move into one file, as other groups of related cases
already are, and the `l1_var_tfvars_module` case drops its second module: a
module that declares no matching variable has nowhere to put a value, so it
cannot fail differently from the one that does.
