--- tests/tfcompat/l1_var_tfvars_convert_test.go
@@ -1,30 +1,0 @@
-// Copyright 2026, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package tfcompat_test
-
-import (
-	"testing"
-
-	"github.com/pulumi-labs/pulumi-hcl/tests/testutil/tfcompat"
-)
-
-// TestL1VarTfvarsConvert pins that a value that does not fit the variable's
-// declared type is an error, not a value that survives unconverted.
-func TestL1VarTfvarsConvert(t *testing.T) {
-	t.Parallel()
-	tfcompat.RunCase(t, "l1_var_tfvars_convert", tfcompat.Case{
-		Stages: []tfcompat.Stage{{ExpectErr: "given value is not suitable for var.n"}},
-	})
-}
