--- tests/tfcompat/l1_var_tfvars_undeclared_test.go
@@ -1,37 +1,0 @@
-// Copyright 2026, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package tfcompat_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	"github.com/pulumi-labs/pulumi-hcl/tests/testutil/tfcompat"
-)
-
-// TestL1VarTfvarsUndeclared pins that a variable-value file setting a name the
-// root module does not declare warns and runs: the value reaches nothing, and
-// its expression is never evaluated.
-func TestL1VarTfvarsUndeclared(t *testing.T) {
-	t.Parallel()
-	tfcompat.RunCase(t, "l1_var_tfvars_undeclared", tfcompat.Case{
-		Stages: []tfcompat.Stage{{
-			AssertOutput: func(t *testing.T, output string) {
-				assert.Contains(t, output, "Value for undeclared variable")
-			},
-		}},
-	})
-}
