+++ tests/tfcompat/testdata/cases/l1_var_tfvars_module/main.tf
@@ -1,17 +1,9 @@
-# `terraform.tfvars` files are loaded automatically for the root module only.
-# The ones under ./declares and ./silent are never read.
+# `terraform.tfvars` files are loaded automatically for the root module only:
+# the one under ./declares sets `who` and is never read.
 module "declares" {
   source = "./declares"
 }
 
-module "silent" {
-  source = "./silent"
-}
-
 output "declares" {
   value = module.declares.who
-}
-
-output "silent" {
-  value = module.silent.who
 }
