--- tests/tfcompat/testdata/cases/l1_var_tfvars_module/silent/main.tf
@@ -1,3 +1,0 @@
-output "who" {
-  value = "silent-literal"
-}
