--- tests/tfcompat/testdata/cases/l1_var_tfvars_module/silent/terraform.tfvars
@@ -1,1 +1,0 @@
-who = "from-module-tfvars"
